requirements.txt
streamlit
pandas
numpy
faiss-cpu
sentence-transformers
python-dotenv
openai